requirements.txt
Flask
flask_apscheduler
backports.zoneinfo;python_version<"3.9"
certifi==2022.5.18.1
charset-normalizer==2.0.12
click==8.1.3
Flask-APScheduler==1.12.3
idna==3.3
importlib-metadata==4.11.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
oauthlib==3.2.0
python-dateutil==2.8.2
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
requests==2.28.0
requests-oauthlib==1.3.1
six==1.16.0
tzdata==2022.1
tzlocal==4.2
urllib3==1.26.9
Werkzeug==2.1.2
zipp==3.8.0
Gunicorn
openai
python-dotenv
